manage errors page and control delete method

+++ app/Http/Controllers/NewsController.php
@@ -104,7 +104,9 @@ class NewsController extends Controller
      */
     public function destroy($id)
     {
+        
         $new = News::find($id);
+        $this->authorize('delete', $new);
         $new->delete();
         return redirect('/mesNews');
     }
